fix_fe_product

+++ backend/database/migrations/2023_11_19_153622_create_carts_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('carts', function (Blueprint $table) {
             $table->id();
-            $table->float('total_price');
+            $table->double('total_price');
             $table->timestamps();
         });
     }
